refactor: move StatelogParser to lib/statelogParser.ts
Pure move + import-path fixes; no behavior change.
